requirements.txt
click
codecov
coverage
cryptography
gradio
imgaug
numpy
onnxruntime
opencv-python-headless
Pillow
pooch
rich
sphinx-click
torch
torchvision
tqdm
wbia-utool
